+++ tests/test_config.py
@@ -20,6 +20,14 @@ run_config:
     start_date: 20200102
     image_pull_secrets: pull1,pull2
     service_account_name: service_account
+    spark:
+      type: dataproc
+      region: europe-west2
+      project_id: sandbox
+      cluster_name: cluster-8084
+      artifacts_path: gs://test/dataproc
+      cluster_config:
+        configBucket: test
     kubernetes_pod_templates:
         custom_template:
             template: |-
@@ -200,6 +208,15 @@ spec:
     name: 1"""
         )
 
+        spark = cfg.run_config.spark
+        assert spark
+        assert spark.type == "dataproc"
+        assert spark.region == "europe-west2"
+        assert spark.project_id == "sandbox"
+        assert spark.cluster_name == "cluster-8084"
+        assert spark.artifacts_path == "gs://test/dataproc"
+        assert spark.cluster_config["configBucket"] == "test"
+
     def test_defaults(self):
         cfg = PluginConfig({"run_config": {}})
 
@@ -211,6 +228,7 @@ spec:
         assert cfg.run_config.start_date is None
         assert cfg.run_config.auth_config.type == "Null"
         assert cfg.run_config.auth_config.params == []
+        assert cfg.run_config.spark
 
         assert cfg.run_config.volume
         assert cfg.run_config.volume.disabled is False